Format fixes, ungate implementation for no_std environment, change docs, add Float types, move to tests mod
